Multilingual-Translator/requirements.txt
fastapi
uvicorn
websockets
python-multipart
langchain
langgraph
groq
python-dotenv
speechrecognition
gtts
pydub